Move `balanceTransactionSpec` to `Cardano.Write.Tx.BalanceSpec`.
