+++ lib/wallet/cardano-wallet.cabal
@@ -921,6 +921,7 @@ test-suite unit
     Cardano.Wallet.Submissions.PrimitivesSpec
     Cardano.Wallet.TokenMetadataSpec
     Cardano.WalletSpec
+    Cardano.Write.Tx.BalanceSpec
     Control.Concurrent.ConciergeSpec
     Control.Monad.UtilSpec
     Data.Function.UtilsSpec
